requirements.txt
crewai>=1.9.0
crewai-tools>=1.9.0
google-generativeai>=0.8.0
python-dotenv
requests
Pillow
instagrapi
